tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/111.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: removing async attribute at runtime</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

        <div id="test"></div>
	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

        <script id="async-script" async src="scripts/include-2.js?pipe=trickle(d1)"></script> 

	<script>
          var t = async_test();

          t.step(function() {
            document.getElementById("async-script").removeAttribute("async");
            var s = document.createElement("script");
            s.async = false;
            s.src = "scripts/include-1.js";
            document.body.appendChild(s);
          });

          addEventListener("load", t.step_func(function () {
            assert_array_equals(eventOrder, ["external script #1", "external script #2"]);
            t.done();
          }), false);
          
        </script>
</body></html>
